+++ frontend/src/Layout.jsx
@@ -3,10 +3,8 @@ import React from 'react';
 export function Layout(props) {
 
     return (
-        <div >
-            <div >
-                {props.children}
-            </div>
+        <div>
+            {props.children}
         </div>
     );
