+++ frontend/src/components/NavBar.css
@@ -1,0 +1,6 @@
+.Appbar {
+    display: flex !important;
+    flex-direction: row !important;
+    justify-content: space-between !important;
+    align-items: center !important;
+}
